refactor CommandBuilder Tests and Builder
